refactor: BetAmount 위치 변경
