+++ src/test/java/domain/money/BetAmountTest.java
@@ -1,4 +1,4 @@
-package view;
+package domain.money;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 
 public class BetAmountTest {
     @Test
-    @DisplayName("베팅 금액은 최소 1000이다")
+    @DisplayName("베팅 금액이 1000 미만이면 예외가 발생한다.")
     void invalidBetAmountTest() {
         assertThatThrownBy(() -> new BetAmount(500))
                 .isInstanceOf(IllegalArgumentException.class)
